Pass Event id as Number instead of String
Seems that router-link always converts the Id to a string. Thus passing it forward to fetchEvent in store will lead to always loading the Event from EventService instead of reusing 'existingEvent'. Actually this bug this has no effect on the UI since loading the Event will always return the same result.