fix: Broken Display of Event Details fixed.
The last commit, which changed the passing of the event ID from string to number, broke the display of newly created events. The reason was that newly created events get a UUID as ID, while the predefined events in db.json had integers as ID. So the problem was in the data, where it has now been fixed by changing the integers to UUIDs and undoing the last change.